Move getBackgroundImageUrl util to lib
