+++ lib/get-background-image-url/getBackgroundImageUrl.spec.js
