+++ lib/get-background-image-url/index.js
@@ -1,0 +1,1 @@
+export { getBackgroundImageUrl } from './getBackgroundImageUrl';
